# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.


    **Configuration environment variables:**
    `INSTANCE_ID` (auto-generated UUID; included in all API response bodies and
    webhook events), `HTTP_ADDR` (:8080), `SIP_BIND_IP`, `SIP_PORT`, `SIP_HOST`,
    `ICE_SERVERS`, `RECORDING_DIR`, `LOG_LEVEL`, `WEBHOOK_URL`,
    `ELEVENLABS_API_KEY`, `VAPI_API_KEY` (VAPI agent provider API key),
    `S3_BUCKET`, `S3_REGION`, `S3_ENDPOINT`, `S3_PREFIX`.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /legs/{id}/play:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: playLeg
      summary: Start audio playback to a leg
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackRequest'
      responses:
        '200':
          description: Playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackResponse'
        '400':
          description: Invalid JSON or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg has no audio writer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rooms/{id}/tts:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: ttsRoom
      summary: Synthesize speech and play it into a room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSRequest'
      responses:
        '200':
          description: TTS playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTSResponse'
        '400':
          description: Invalid JSON, missing text/voice, or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: Room has no participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No API key provided for the selected provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    LegId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Leg ID
    RoomId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Room ID
...
  responses:
    LegNotFound:
      description: Leg not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        error:
          type: string
          description: Error message
      required: [error]
...
    PlaybackRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: URL of the audio file (mutually exclusive with tone)
        tone:
          type: string
          description: >
            Built-in telephone tone name. Format: {country}_{type} or bare {type} (defaults to US).
            Types: ringback, busy, dial, congestion.
            Countries: us, gb, de, fr, au, jp, it, in, br, pl, ru.
            Examples: us_ringback, gb_busy, dial.
        mime_type:
          type: string
          description: MIME type (e.g. audio/wav). Required when using url.
        repeat:
          type: integer
          description: Number of times to repeat playback (url only)
          default: 0
        volume:
          type: integer
          description: Volume adjustment in dB (-8 to 8)
          minimum: -8
          maximum: 8
          default: 0
      oneOf:
        - required: [url, mime_type]
        - required: [tone]
    TTSRequest:
      type: object
      properties:
        text:
          type: string
          description: Text to synthesize
        voice:
          type: string
          description: >
            Provider-specific voice identifier. ElevenLabs: voice name or ID.
            AWS Polly: voice ID (e.g. Joanna, Matthew). Google Cloud: voice name —
            either full format (e.g. en-US-Neural2-F) or short name for Gemini
            models (e.g. Achernar, Kore).
        provider:
          type: string
          enum: [elevenlabs, aws, google]
          description: 'TTS provider: "elevenlabs" (default), "aws", or "google"'
        model_id:
          type: string
          description: >
            Provider-specific model/engine. ElevenLabs: model ID.
            AWS Polly: engine (standard, neural, long-form, generative; default neural).
            Google Cloud: model name (e.g. gemini-2.5-pro-tts, chirp3-hd).
        language:
          type: string
          description: >
            Language code (e.g. "en-US", "pl-pl"). Required for Google Gemini TTS
            voices that use short names (e.g. Achernar). Auto-extracted from full
            voice names like en-US-Neural2-F.
        prompt:
          type: string
          description: >
            Style/tone instruction for promptable voice models (Google Gemini TTS only).
            E.g. "Read aloud in a warm, welcoming tone."
        volume:
          type: integer
          description: Volume adjustment in dB (-8 to 8)
          minimum: -8
          maximum: 8
          default: 0
        api_key:
          type: string
          description: >
            ElevenLabs: API key override (falls back to ELEVENLABS_API_KEY env var).
            AWS: optional ACCESS_KEY:SECRET_KEY override (falls back to default AWS
            credential chain). Google Cloud: optional API key override (falls back
            to Application Default Credentials).
      required: [text, voice]
...
    PlaybackResponse:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        playback_id:
          type: string
          description: Unique playback identifier
        status:
          type: string
          example: playing
      required: [playback_id, status]
    TTSResponse:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        tts_id:
          type: string
          description: Unique TTS playback identifier
        status:
          type: string
          example: playing
      required: [tts_id, status]
...